+++ include/fastfields/api/cpu/reg_flow.h
@@ -41,7 +41,7 @@ void flow_matvec(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -59,7 +59,7 @@ void flow_addmatvec_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -77,7 +77,7 @@ void flow_submatvec_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -94,7 +94,7 @@ void flow_diag(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 
@@ -116,7 +116,7 @@ void flow_adddiag_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -137,7 +137,7 @@ void flow_subdiag_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -178,7 +178,7 @@ void flow_kernel(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -199,7 +199,7 @@ void flow_addkernel_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -220,7 +220,7 @@ void flow_subkernel_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -250,7 +250,7 @@ void flow_relax(
           int8_t     bound     = 0,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -276,7 +276,7 @@ void flow_forward(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -301,7 +301,7 @@ void flow_precond(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -322,7 +322,7 @@ void flow_precond_(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -354,7 +354,7 @@ void flow_matvec_rls(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -372,7 +372,7 @@ void flow_diag_rls(
           double     div       = 0.0,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -399,7 +399,7 @@ void flow_relax_rls(
           int8_t     bound     = 0,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 } // namespace cpu
