+++ include/fastfields/api/cpu/resize.h
@@ -33,7 +33,7 @@ void resample(
           double     shift  ,
     const double   * scale  ,
           int        ndim   ,
-          int        stream = 0
+          intptr_t   stream = 0
 );
 
 } // namespace cpu
